Done 88%

+++ WebBanTraSua/Models/EF/HoaDon.cs
@@ -51,7 +51,7 @@
         public int? soLuong { get; set; }
 
         [Display(Name = "Tổng tiền")]
-        public int? tongTien { get; set; }
+        public double? tongTien { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<ChiTietHoaDon> ChiTietHoaDons { get; set; }
